moves remaining types into new lib
