--- apps/menu-matriarch/src/app/shared/models/tag.interface.ts
@@ -1,3 +1,0 @@
-import { TagDto } from './dtos/tag-dto.interface';
-
-export type Tag = TagDto;
