--- apps/menu-matriarch/src/app/shared/models/user.interface.ts
@@ -1,3 +1,0 @@
-import { UserDto } from './dtos/user-dto.interface';
-
-export type User = UserDto;
