+++ libs/menu-matriarch/types/src/index.ts
@@ -1,3 +1,23 @@
+export * from './lib/dtos/dish-dto.interface';
+export * from './lib/dtos/meal-dto.interface';
+export * from './lib/dtos/menu-dto.interface';
+export * from './lib/dtos/tag-dto.interface';
+export * from './lib/dtos/user-dto.interface';
+
+export * from './lib/entities/dish.interface';
+export * from './lib/entities/meal.interface';
+export * from './lib/entities/menu.interface';
+export * from './lib/entities/tag.interface';
+export * from './lib/entities/user.interface';
+
+export * from './lib/date-name-display.type';
+export * from './lib/day.type';
+export * from './lib/dish-type.type';
 export * from './lib/endpoint.enum';
+export * from './lib/filtered-dishes.interface';
+export * from './lib/menu-entry.interface';
+export * from './lib/orientation.type';
 export * from './lib/planner-view.type';
 export * from './lib/route.enum';
+export * from './lib/tag-model.interface';
+export * from './lib/user-preferences.interface';
