+++ libs/menu-matriarch/types/src/lib/day.type.ts
