+++ libs/menu-matriarch/types/src/lib/entities/dish.interface.ts
@@ -1,4 +1,4 @@
-import { DishDto } from './dtos/dish-dto.interface';
+import { DishDto } from '../dtos/dish-dto.interface';
 import { Tag } from './tag.interface';
 
 export interface Dish extends Omit<DishDto, 'ingredientIds' | 'tagIds'> {
