+++ libs/menu-matriarch/types/src/lib/entities/tag.interface.ts
@@ -1,0 +1,3 @@
+import { TagDto } from '../dtos/tag-dto.interface';
+
+export type Tag = TagDto;
