+++ libs/menu-matriarch/types/src/lib/filtered-dishes.interface.ts
@@ -1,5 +1,5 @@
 import { DishType } from './dish-type.type';
-import { Dish } from './dish.interface';
+import { Dish } from './entities/dish.interface';
 
 export interface FilteredDishesGroup {
   type: DishType;
