+++ libs/menu-matriarch/types/src/lib/tag-model.interface.ts
@@ -1,4 +1,4 @@
-import { Tag } from './tag.interface';
+import { Tag } from './entities/tag.interface';
 
 export interface TagModel extends Tag {
   checked: boolean;
